~ fixed Role column being the wrong type on the account's table

+++ Database/Server/account.sql
@@ -31,7 +31,7 @@ CREATE TABLE IF NOT EXISTS `account` (
   `accountID` int(10) unsigned NOT NULL AUTO_INCREMENT,
   `accountName` varchar(48) CHARACTER SET utf8 NOT NULL,
   `password` blob NOT NULL,
-  `role` int(10) unsigned NOT NULL,
+  `role` bigint(20) unsigned NOT NULL,
   `online` tinyint(1) NOT NULL,
   `banned` tinyint(1) NOT NULL,
   PRIMARY KEY (`accountID`),
